requirements.txt
dearpygui
python-can
pluggy
numpy
scipy